Add files via upload
Python implementation of the KBaNN for the Forrester test function. 

Knowledge Based Neural Networks are a bi-fidelity machine learning architecture that allow the outputs of a coarse scale model, Fc(x), to be augmented by the predictions of a neural network. Having been trained using a dataset comprising outputs of a high-fidelity model, Fe(x), the KBaNN corrects the outputs of the coarse model to emulate the output of Fe(x). The architecture is based on the KBaNN proposed in Wang et al (1997) but adapted for bi-fidelity modelling. The formalism has also been modified to produce an additive rather than multiplicative correction to . It also now incorporates  L2 regularisation. The KBaNN formulation attempots to leverage the flexibility and scalability of neural networks, while addressing the criticism that they are data-driven black boxes by incorporating the coarse model. 

Please do provide us with feedback on the code, we are happy to engage to improve it! 

Mathematical details may be found in Pepper, Gaymann, Montomoli, and Sharma, "Local Navier-Stokes approximation with a Knowledge Based Neural Network" (2021)